--- src/pages/Profile/Body/GeneralInfo/Content/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Content";
