+++ src/pages/Profile/Body/Program/Milestones.tsx
@@ -1,0 +1,3 @@
+export default function Milestones() {
+  return <div>milestones</div>;
+}
