+++ src/pages/Profile/Body/Program/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Program";
